Now builds on Linux again, plus added support for Qt5 and OS/2.
It seems the maintainer dropped support for any other platform but
Windows while implementing features in his ToDo list for release 3.0:

 http://lookias.worms2d.info/forum/viewtopic.php?f=9&t=53

So I had to add ifdefs for the WormNAT2 support feature, which seems to
be Windows-only, and also I had to add #ifdef Q_WS_WIN in some places in
order to get it built.

Now, with WormNAT2 support turned off (undef WITH_WORMNAT_SUPPORT), it
seems to be able to build on Linux.

Also, to build TWS with Qt5 (or for OS/2), you need to turn off the
support for phonon (deprecated sound interfaces). I've added the proper
ifdefs in the right places in this commit.
It will be turned off automatically if building with Qt5, or if the
build target is OS/2. To force the support off manually, edit the Qt
project file netcoupler.pro.