Solution to the issue #15 by ZexorZ.
TWS has a feature to display clan information including filename and the URL
of its site the user can visit by clicking on the clan field in the userlist.

Back then Lookie's clan db was used which contained only two clans. With this
commit TUS clan database (http://www.tus-wa.com/groups/) is getting used.

ZexorZ has written a parser script for that web page (included in the repo as
stuff/tus-wa-groups-db-parser.php with this commit) in PHP and hosted it
on his server (http://ultimateclan.net/clanparse.php). The script returns
the parsed TUS clan database in a compatible with TWS format.

Big thanks to him.