More menus now work; UI and About improvements; see commit msg.
Lots of margins in the main form are now set to 0 to save space.
The Lock UI feature now works and causes the docks and the toolbar
to lock in place, also theirs title bar is replaced with an empty
QWidget (when the "UI" is "locked"), and returned back on "unlock".

The text of the About box is now set up by the res/about.html file
(see about.cpp). The text is rewritten also.

Other UI changes maybe.